Move themes.ts file to shared helper folder
