+++ packages/devextreme/testing/testcafe/helpers/themes.ts
