+++ packages/devextreme/testing/testcafe/tests/dataGrid/focus/focusedRow/markup.ts
@@ -3,7 +3,7 @@ import { changeTheme } from '../../../../helpers/changeTheme';
 import createWidget from '../../../../helpers/createWidget';
 import url from '../../../../helpers/getPageUrl';
 import DataGrid from '../../../../model/dataGrid';
-import { Themes } from '../../helpers/themes';
+import { Themes } from '../../../../helpers/themes';
 
 fixture.disablePageReloads`Focused row - markup`
   .page(url(__dirname, '../../../container.html'));
